And the functionality starts.
Started work on the team commands. All are in place. Allows the user (admin) to edit the different pieces for each team.

Base functionalit has now been added all over. Next steps to work on are the "scheduling" parts. This includes making the weekly messages that the users cna react to.

After that, maybe start work on arranging "events" that people can sign up to but ONLY if they are on THAT team. Maybe add a "versus" option that "syncs" the options between multiple teams.